cell_type & testcases
