+++ go.mod
@@ -2,10 +2,7 @@ module github.com/jmnote/nbformat
 
 go 1.22.3
 
-require (
-	github.com/stretchr/testify v1.9.0
-	k8s.io/utils v0.0.0-20240502163921-fe8a2dddb1d0
-)
+require github.com/stretchr/testify v1.9.0
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
