+++ tests/README.md
@@ -1,0 +1,4 @@
+Sources of ipynb files:
+- myfiles - ipynb files generated using `nbconvert`
+- nbconvert - ipynb files from https://github.com/jupyter/nbconvert/tree/v7.16.4/tests
+- nbformat - ipynb files (excluding invalid, v2, and v3 files) from https://github.com/jupyter/nbformat/tree/v5.10.4/tests
