--- tests/nbconvertfiles/README.md
@@ -1,1 +1,0 @@
-ipynb files from https://github.com/jupyter/nbconvert/tree/v7.16.4/tests
